+++ src/features/Movies/MovieList/Success/index.js
@@ -1,5 +1,5 @@
 import { ListItem, Item } from "./styled";
-import Tile from "../../../../common/Tile";
+import Tile from "../../../../common/Tiles/Tile";
 import Section from "../../../../common/Section";
 import { Main } from "../../../../common/main";
 
@@ -13,11 +13,11 @@ const Success = ({ results }) => (
             ? results.map((movie) => (
                 <Item key={movie.id}>
                   <Tile
-                    title={movie.original_title}
+                    name={movie.original_title}
                     year={movie.release_date.substring(0, 4)}
-                    rating={movie.vote_average.toFixed(1)}
+                    score={movie.vote_average.toFixed(1)}
                     votes={movie.vote_count}
-                    poster={movie.poster_path}
+                    image={movie.poster_path}
                     genres={"."}
                     id={movie.id}
                   />
